requirements.txt
alabaster==0.7.13
attrs==23.1.0
Babel==2.13.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
defusedxml==0.7.1
docutils==0.18.1
EasyProcess==1.1
entrypoint2==1.1
evdev==1.6.1
fastjsonschema==2.18.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.10.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mistune==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nbsphinx==0.9.3
networkx==3.2
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
Pillow==10.1.0
platformdirs==3.11.0
psutil==5.9.6
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.0
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
supervision==0.16.0
sympy==1.12
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.200
urllib3==2.0.7
webencodings==0.5.1
zipp==3.17.0
